AI-News-Research-Agent/requirements.txt
openai==1.59.3
tavily-python==0.5.0
python-dotenv==1.0.1
pydantic==2.10.4
streamlit==1.32.0
